Front

See merge request s09-webmobile1-sub2/S09P12E201!73

Former-commit-id: d4ce2b04188dd9b32d6832a5899f950973cdd63f

+++ backend/src/main/java/com/ssafy/api/service/TrophyRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.ssafy.api.model.Trophy;
 
 @Repository
-public interface TrophyRepository extends JpaRepository<Trophy, Integer>{
+public interface TrophyRepository extends JpaRepository<Trophy, Long>{
 
 }
